requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20210324.1=h9c3ff4c_0
albumentations=0.4.3=pypi_0
arrow-cpp=4.0.0=py38h8e6f228_3_cpu
attrs=21.4.0=pyhd3eb1b0_0
autograd=1.3=pyhd3eb1b0_1
autograd-gamma=0.4.1=py_0
aws-c-cal=0.5.11=h95a6274_0
aws-c-common=0.6.2=h7f98852_0
aws-c-event-stream=0.2.7=h3541f99_13
aws-c-io=0.10.5=hfb6a706_0
aws-checksums=0.1.11=ha31a3da_7
aws-sdk-cpp=1.8.186=hb4091e7_3
bcj-cffi=0.5.1=py38h295c915_0
blas=1.0=mkl
bottleneck=1.3.5=py38h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotli-python=1.0.9=py38h6a678d5_7
brotlicffi=1.0.9.2=py38h295c915_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.07.19=h06a4308_0
cairo=1.14.12=h80bd089_1005
certifi=2022.9.14=py38h06a4308_0
cffi=1.15.0=py38h7f8727e_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
clip=1.0=pypi_0
cryptography=37.0.1=py38h9ce1e76_0
cudatoolkit=10.2.89=hfd86e86_1
cvxopt=1.2.5=py38h9f755b6_0
cvxpy=1.0.31=py38_0
cvxpy-base=1.0.31=py38hf484d3e_0
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
dill=0.3.4=pyhd3eb1b0_0
ecos=2.0.7.post1=py38h6323ea4_0
eli5=0.10.1=py38_1
et_xmlfile=1.1.0=py38h06a4308_0
expat=2.2.10=he6710b0_2
fastcache=1.1.0=py38h7b6447c_0
feather-format=0.4.1=pyh9f0ad1d_0
fontconfig=2.13.1=he4413a7_1000
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
ftfy=6.1.1=pypi_0
future=0.18.2=py38_1
gettext=0.21.0=hf68c758_0
gflags=2.2.2=he6710b0_0
giflib=5.2.1=h7b6447c_0
glib=2.56.2=hd408876_0
glog=0.5.0=h2531618_0
glpk=4.65=h276157c_3
gmp=6.2.1=h295c915_3
graphite2=1.3.14=h295c915_1
graphviz=2.38.0=hcf1ce16_1009
grpc-cpp=1.37.1=h2519f57_2
gsl=2.4=h14c3975_4
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=3.7.0=py38h737f45e_0
harfbuzz=1.9.0=he243708_1001
hdf5=1.10.6=hb1b8bf9_0
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
imgaug=0.2.6=pypi_0
importlib-metadata=4.11.3=py38h06a4308_0
importlib_metadata=4.11.3=hd3eb1b0_0
intel-openmp=2022.1.0=h9e868ea_3769
jdcal=1.4.1=pyhd3eb1b0_0
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
kaggle=1.5.12=pypi_0
keras=2.6.0=pyhd8ed1ab_0
kiwisolver=1.4.2=py38h295c915_0
krb5=1.19.2=hac12032_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcurl=7.84.0=h91b91d3_0
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libevent=2.1.10=hcdb4288_3
libffi=3.2.1=he1b5a44_1007
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1234567_1
libllvm11=11.1.0=h9e868ea_5
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.16.0=h780b84a_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libthrift=0.14.1=he6d91bd_2
libtiff=4.1.0=hc7e4089_6
libtool=2.4.6=h295c915_1008
libutf8proc=2.6.1=h27cfd23_0
libuuid=2.32.1=h14c3975_1000
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.1.0=2
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
lifelines=0.24.2=py_0
llvmlite=0.38.0=py38h4ff587b_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.1.1=py38h7f8727e_0
matplotlib=3.5.2=py38h06a4308_0
matplotlib-base=3.5.2=py38hf590b9c_0
metis=5.1.0=hf484d3e_4
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.3.0=py38h54f3939_0
mkl_random=1.1.1=py38h0573a6f_0
multiprocess=0.70.12.2=py38h7f8727e_0
multivolumefile=0.2.3=pyhd3eb1b0_0
munkres=1.1.4=py_0
ncurses=6.3=h5eee18b_3
nibabel=3.0.2=py_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numba=0.55.1=py38h51133e4_0
numexpr=2.7.3=py38hb2eb853_0
numpy=1.17.4=pypi_0
numpy-base=1.19.2=py38hfa32c7d_0
opencv-python=4.2.0.34=pypi_0
openpyxl=3.0.5=py_0
openssl=1.1.1q=h7f8727e_0
orc=1.6.7=h89a63ab_2
osqp=0.5.0=py38hb38ba16_3
packaging=21.3=pyhd3eb1b0_0
pandas=0.25.3=pypi_0
pango=1.40.14=hf0c64fd_1003
parquet-cpp=1.5.1=h34088ae_4
pcre=8.45=h295c915_0
pillow=8.1.1=pypi_0
pip=22.1.2=py38h06a4308_0
pixman=0.34.0=h14c3975_1003
py7zr=0.16.1=pyhd3eb1b0_1
pyarrow=4.0.0=py38hc9229eb_3_cpu
pycox=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodomex=3.15.0=py38h5eee18b_0
pydicom=1.4.2=py_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py38h06a4308_0
pypng=0.0.20=py_0
pyppmd=0.16.1=py38h295c915_0
pyqt=5.9.2=py38h05f1152_4
pysocks=1.7.1=py38h06a4308_0
python=3.8.2=h9d8adfe_4_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-graphviz=0.20.1=py38h06a4308_0
python-slugify=6.1.2=pypi_0
python_abi=3.8=1_cp38
pytorch=1.9.0=py3.8_cuda10.2_cudnn7.6.5_0
pytz=2022.1=py38h06a4308_0
pyyaml=6.0=pypi_0
pyzstd=0.14.4=py38hf3a4ac8_2
qt=5.9.7=h5867ecd_1
re2=2021.04.01=h9c3ff4c_0
readline=8.1.2=h7f8727e_1
regex=2022.8.17=pypi_0
requests=2.28.1=py38h06a4308_0
s2n=1.0.10=h9b69904_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.22.2.post1=pypi_0
scikit-survival=0.14.0=py38h1211aff_0
scipy=1.6.2=py38h91f5cce_0
scs=2.1.2=py38hb38ba16_0
seaborn=0.11.0=py_0
setuptools=63.4.1=py38h06a4308_0
simpleitk=2.0rc2=py38hf484d3e_0
singledispatch=3.7.0=pyhd3eb1b0_1001
sip=4.19.13=py38h295c915_0
six=1.16.0=pyhd3eb1b0_1
sklearn-pandas=2.2.0=pyhd8ed1ab_0
snappy=1.1.9=h295c915_0
sqlite=3.39.2=h5082296_0
suitesparse=5.2.0=h9e4a6bb_0
tabulate=0.8.10=py38h06a4308_0
tbb=2021.5.0=hd09550d_0
tensorboard=2.2.2=pypi_0
tensorflow=2.2.1=pypi_0
tensorflow-estimator=2.2.0=pypi_0
text-unidecode=1.3=pypi_0
texttable=1.6.4=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
torchtuples=0.2.0=py_0
torchvision=0.2.2=py_3
tornado=6.2=py38h5eee18b_0
tqdm=4.45.0=pypi_0
typing_extensions=4.3.0=py38h06a4308_0
urllib3=1.26.11=py38h06a4308_0
wcwidth=0.2.5=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xlrd=1.2.0=py_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxpm=3.5.13=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-libxt=1.1.5=h516909a_1003
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h27cfd23_1007
xz=5.2.5=h7f8727e_1
zipp=3.8.0=py38h06a4308_0
zlib=1.2.12=h5eee18b_3
zstd=1.4.9=haebb681_0
